+++ src/jiratui/commands/handler.py
@@ -12,6 +12,8 @@ from jiratui.exceptions import (
 from jiratui.models import (
     IssueComment,
     IssueTransition,
+    JiraBaseIssue,
+    JiraIssue,
     JiraIssueSearchResponse,
     JiraUser,
     JiraUserGroup,
@@ -383,15 +385,21 @@ class CommandHandler:
             CLIException: if an error occurs while getting the comment(s).
         """
 
-        response: APIControllerResponse = asyncio.run(
-            self.api.get_issue(issue_id_or_key=key, fields=fields)
-        )
-        if response.success:
-            return response.result
-        raise CLIException(
-            'An error occurred while trying to retrieve the work item.',
-            extra={'work_item_key': key, 'error_message': response.error},
-        )
+        try:
+            response: APIControllerResponse = asyncio.run(
+                self.api.get_issue(issue_id_or_key=key, fields=fields)
+            )
+        except Exception as e:
+            raise CLIException(
+                'An error occurred while trying to retrieve the work item.',
+                extra={'work_item_key': key, 'error_message': str(e)},
+            ) from e
+        if not response.success or not response.result:
+            raise CLIException(
+                'An error occurred while trying to retrieve the work item.',
+                extra={'work_item_key': key, 'error_message': response.error},
+            )
+        return response.result
 
     async def get_metadata(self, key: str, field_id: str | None = None) -> dict:
         """Retrieves the metadata of a work item.
@@ -500,3 +508,125 @@ class CommandHandler:
                 },
             )
         return {'metadata': response.result}
+
+    async def clone_work_item(
+        self, work_item_key: str, summary: str | None = None, clone_status: bool = False
+    ) -> dict:
+        """Clones a work item.
+
+        The fields that are cloned are the following:
+            - summary
+            - description
+            - project key
+            - issue type
+            - priority
+            - due date
+            - assignee
+            - reporter
+            - parent
+
+        Args:
+            work_item_key: the key of the item we want to clone.
+            summary: the summary (aka. title) of the new work item. Defaults to: "(clone) <source-item-summary>"
+            clone_status: when True the new item will have the same status as the source item. The default is False.
+
+        Returns:
+            A dictionary with a key `work_item` with the cloned work item and, a key `key` with the key of the new item.
+
+        Raises:
+            CLIException: if the (source) work item can not be found.
+            CLIException: if unable to clone the work item.
+            CLIException: if an error occurs while trying to retrieve the (cloned) work item.
+        """
+
+        get_item_response: APIControllerResponse = await self.api.get_issue(
+            issue_id_or_key=work_item_key
+        )
+        items_found: JiraIssueSearchResponse | None
+        if not get_item_response.success or not (items_found := get_item_response.result):
+            raise CLIException(
+                f'Unable to find the work item with key {work_item_key}',
+                extra={'work_item_key': work_item_key, 'error_message': get_item_response.error},
+            )
+
+        issue: JiraIssue = items_found.issues[0]
+
+        # build the payload for creating the clone
+        clone_data: dict = {
+            'summary': summary or f'(cloned) {issue.summary}',
+            'project_key': issue.project.key,
+            'issue_type_id': issue.issue_type.id,
+            'description': issue.description,
+        }
+
+        if issue.assignee:
+            clone_data['assignee_account_id'] = issue.assignee.account_id
+        if issue.reporter:
+            clone_data['reporter_account_id'] = issue.reporter.account_id
+        if issue.parent_key:
+            clone_data['parent_key'] = issue.parent_key
+        if issue.priority:
+            clone_data['priority'] = issue.priority.id
+        if issue.due_date:
+            clone_data['duedate'] = issue.due_date.strftime('%Y-%m-%d')
+
+        create_work_item_response: APIControllerResponse = await self.api.create_work_item(
+            clone_data
+        )
+
+        if not create_work_item_response.success or not create_work_item_response.result:
+            raise CLIException(
+                'Unable to clone the work item.',
+                extra={
+                    'work_item_key': work_item_key,
+                    'error_message': create_work_item_response.error,
+                },
+            )
+
+        cloned_item: JiraBaseIssue | None = create_work_item_response.result
+
+        # if required then transition the status of the new item to match the status of the source item
+        if clone_status:
+            await self.api.transition_issue_status(cloned_item.key, issue.status.id)
+
+        # retrieve the basic details of the new item
+        try:
+            get_issue_response: APIControllerResponse = await self.api.get_issue(
+                issue_id_or_key=cloned_item.key
+            )
+        except Exception as e:
+            raise CLIException(
+                f'An error occurred while trying to retrieve the cloned item with key {cloned_item.key}',
+                extra={'work_item_key': cloned_item.key, 'error_message': str(e)},
+            ) from e
+        if not get_issue_response.success or not get_issue_response.result:
+            raise CLIException(
+                f'Not Found: (cloned) item with key {cloned_item.key}',
+                extra={'work_item_key': cloned_item.key, 'error_message': get_issue_response.error},
+            )
+
+        return {
+            'key': cloned_item.key,
+            'work_item': get_issue_response.result,
+        }
+
+    async def delete_work_item(self, work_item_key: str) -> bool:
+        """Deletes a work item.
+
+        Args:
+            work_item_key: the key of the item we want to delete.
+
+        Returns:
+            True if the deletion was successful; a CLIException otherwise.
+
+        Raises:
+            CLIException: if an error occurred while trying to delete the work item.
+        """
+
+        response: APIControllerResponse = await self.api.delete_work_item(work_item_key)
+        if not response.success:
+            raise CLIException(
+                'An error occurred while trying to delete the work item.',
+                extra={'work_item_key': work_item_key, 'error_message': response.error},
+            )
+        return True